Migrate functional tests over haskell-lsp-test
