+++ haskell-ide-engine.cabal
@@ -187,6 +187,7 @@ test-suite haskell-ide-func-test
   hs-source-dirs:      test
   main-is:             Functional.hs
   other-modules:       TestUtils
+                       FunctionalDispatch
   build-depends:       aeson
                      , base
                      , containers
@@ -194,10 +195,12 @@ test-suite haskell-ide-func-test
                      , filepath
                      , ghc-mod-core
                      , haskell-lsp
+                     , haskell-lsp-test
                      , haskell-ide-engine
                      , hie-plugin-api
                      , hslogger
                      , hspec
+                     , lens
                      , stm
                      , text
                      , unordered-containers
@@ -207,6 +210,7 @@ test-suite haskell-ide-func-test
      ghc-options:      -Werror
   default-language:    Haskell2010
   build-tool-depends:  hspec-discover:hspec-discover
+
 
 source-repository head
   type:     git
