+++ stack.yaml
@@ -39,6 +39,8 @@ extra-deps:
 # - haskell-lsp-types-0.2.2.0
 - syz-0.2.0.0
 - temporary-1.2.1.1
+- github: Bubba/haskell-lsp-test
+  commit: 88c70a40654c7152fb50b2a4e171fbdc00324f51
 
 flags:
   haskell-ide-engine:
